# Conflicts:
#	server/src/main/java/com/tripco/t13/server/Distance.java

+++ server/src/main/java/com/tripco/t13/server/Calculate.java
@@ -23,7 +23,7 @@ public class Calculate {
 
         distance = gson.fromJson(requestBody, Distance.class);
 
-        int radius = distance.getRadius(distance.units);
+        float radius = distance.getRadius(distance.units);
         //Perform the Calculation
         distance.distance = distance.getDistanceNum(distance.origin.latitude,distance.origin.longitude,distance.destination.latitude,distance.destination.longitude, radius); //Added 2 in place of paramaters
